WORK

+++ controller/ProductController.js
@@ -3,6 +3,7 @@ const fs = require('fs')
 const { ObjectId } = require('mongodb')
 
 const ProductModel = require('../model/ProductModel')
+const CategoryModel = require('../model/CategoryModel')
 
 const getCurrentIndex = (articul, index, curentImages) => {
   if (!curentImages.find(el => el.includes(`${articul}_${index}`))) {
@@ -15,7 +16,7 @@ const getCurrentIndex = (articul, index, curentImages) => {
 class ProductController {
   async index(req, res) {
     try {
-      const limit = Number(req.query.limit) || 25
+      const limit = Number(req.query.limit) || 12
       const payload = {
         visibility: true
       }
@@ -24,21 +25,32 @@ class ProductController {
         payload.gender = ObjectId(req.query.gender)
       }
       if (req.query.category) {
-        payload.category = ObjectId(req.query.category)
+        const category = await CategoryModel.findById(req.query.category)
+        if (!category.parent) {
+          const categorys = await CategoryModel.find({ parent: ObjectId(category._id) })
+          payload.category = { $in: [] }
+          categorys.forEach(el => {
+            payload.category.$in.push(ObjectId(el._id))
+          })
+        } else {
+          payload.category = ObjectId(category._id)
+        }
+      }
+      if (req.query.size) {
+        payload.size = { $in: req.query.size }
       }
 
       const products = await ProductModel.aggregate([
-        { $sort: { dateIn: -1 } },
         { $match: { ...payload } },
-        { $group: { _id: "$articul", products: { $push: "$$ROOT" } } }
+        { $group: { _id: "$articul", name: { $first: "$name" }, dateIn: { $first: "$dateIn"}, products: { $push: "$$ROOT" } } },
+        { $sort: { name: 1 } }
       ]).limit(limit)
 
-      const images = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(product.articul))
       const productWithImage = products.map(product => {
-        const curentImages = images.filter(image => image.includes(product._id))
-        return product = {
+        const images = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(product._id))
+        return {
           ...product,
-          images: curentImages
+          images: images
         }
       })
       return res.status(200).json({ products: productWithImage })
@@ -50,30 +62,88 @@ class ProductController {
 
   async adminIndex(req, res) {
     try {
-      const products = await ProductModel.find().sort({dateIn: -1})
-        .populate('gender')
+      const page = req.query.page
+      const limit = Number(req.query.limit) || 40
+      const offSet = limit * page - limit
+      const payload = {
+        visibility: true
+      }
+      if (req.query.search) {
+        payload.$or = [
+          { codeProduct:  new RegExp(req.query.search, 'i')},
+          { codeBox:  new RegExp(req.query.search, 'i')},
+          { articul:  new RegExp(req.query.search, 'i')}
+        ]
+      }
+
+      const products = await ProductModel.find({...payload})
         .populate('category')
         .populate('pair')
         .populate('seller')
+        .sort({'dateIn': -1})
+        .skip(offSet)
+        .limit(limit)
 
-      return res.status(200).json({ products })
+      const totalProducts = await ProductModel.countDocuments({...payload})
+      
+      const productWithImage = products.map((product) => {
+        if (product.articul) {
+          const images = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(product.articul))
+          return {
+            ...product._doc,
+            images: images
+          }
+        } else {
+          return {
+            ...product._doc,
+            images: []
+          }
+        }
+      })
+
+      return res.status(200).json({ 
+        products: productWithImage,
+        totalCount: totalProducts,
+        totalPages: Math.ceil(totalProducts / limit)
+      })
     } catch (e) {
       console.log(e);
       res.status(500).json({ message: e.message })
     }
   }
 
+  async adminSearchByArticul(req, res) {
+    try {
+      const payload = {}
+      const id = req.params.id
+      
+      payload.articul = new RegExp(id, 'i')
+
+      const product = await ProductModel.findOne({...payload})
+      
+      let productWithImage = {}
+      
+      if (product.articul) {
+        const images = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(product.articul))
+        productWithImage = {
+          ...product._doc,
+          images: images
+        }
+      }
+
+      return res.status(200).json({ product: productWithImage })
+    } catch (e) {
+      console.log(e);
+      res.status(500).json({ message: e.message })
+    } 
+  }
+
   async adminCreate(req, res) {
     try {
       const body = req.body
 
-      const { images } = req.files || {}
-
-      if (images?.length) {
-        images.forEach((image, index) => {
-          const imageName = `${articul}_${index}.jpg`
-          image.mv(path.resolve(__dirname, '..', 'static', imageName))
-        })
+      if (!body.pair) {
+        delete body.pair
       }
 
       const newProduct = new ProductModel(body)
@@ -97,9 +167,14 @@ class ProductController {
         .populate('category')
         .populate('pair')
         .populate('seller')
+
+      const productSizes = await ProductModel.aggregate([
+          { $match: { articul: product.articul, visibility: true } },
+          { $group: { _id: "$size" } },
+        ])
       const curentImages = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(product.articul))
       if (product !== null) {
-        return res.status(200).json({ product: { ...product, images: curentImages } })
+        return res.status(200).json({ product: { ...product._doc, images: curentImages, size: productSizes.map((el) => el._id) } })
       } else {
         return res.status(500).json({ message: 'Товар не найден'})
       }
@@ -113,13 +188,10 @@ class ProductController {
     try {
       const id = req.params.id
       const product = await ProductModel.findById(id)
-        .populate('gender')
-        .populate('category')
-        .populate('pair')
-        .populate('seller')
+
       const curentImages = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(product.articul))
       if (product !== null) {
-        return res.status(200).json({ product: { ...product, images: curentImages } })
+        return res.status(200).json({ product: { ...product._doc, images: curentImages } })
       } else {
         return res.status(500).json({ message: 'Товар не найден'})
       }
@@ -134,30 +206,6 @@ class ProductController {
       let fileChange = false
       const id = req.params.id
       const updateProduct = req.body
-      delete updateProduct.deleteImages
-
-      const deleteImages = req.body.deleteImages
-      const { images } = req.files || {}
-
-      if (deleteImages?.length) {
-        const existFile = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(test.articul))
-        deleteImages.forEach((deleteImage) => {
-          if (deleteImage && existFile.includes(deleteImage)) {
-            fs.unlinkSync(path.resolve(__dirname, '..', 'static', deleteImage))
-          }
-        })
-        fileChange = true
-      }
-
-      if (images?.length) {
-        images.forEach((image) => {
-          const curentImages = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(test.articul))
-          const curentIndex = getCurrentIndex(test.articul, 0, curentImages)
-          const imageName = `${test.articul}_${curentIndex}.jpg`
-          image.mv(path.resolve(__dirname, '..', 'static', imageName))
-        })
-        fileChange = true
-      }
 
       const updatedProduct = await ProductModel.updateOne(
         { _id: id }, 
@@ -183,23 +231,27 @@ class ProductController {
     try {
       const id = req.params.id
       const product = await ProductModel.findById(id)
-      const products = await ProductModel.find({ articul: product.articul, visibility: true })
-
-      if (products.length === 1) {
-        const existFile = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(product.articul))
-        existFile.forEach((file) => {
-          if (file) {
-            fs.unlinkSync(path.resolve(__dirname, '..', 'static', file))
-          }
-        })
-      }
-
-      const deletedProduct = await ProductModel.remove({ _id: id })
-
-      if (deletedProduct.deletedCount !== 0) {
-        return res.status(200)
+      if (product) {
+        const products = await ProductModel.find({ articul: product.articul, visibility: true })
+  
+        if (products.length === 1) {
+          const existFile = fs.readdirSync(path.resolve(__dirname, '..', 'static')).filter(el => el.includes(product.articul))
+          existFile.forEach((file) => {
+            if (file) {
+              fs.unlinkSync(path.resolve(__dirname, '..', 'static', file))
+            }
+          })
+        }
+  
+        const deletedProduct = await ProductModel.remove({ _id: id })
+  
+        if (deletedProduct.deletedCount !== 0) {
+          return res.status(200).json({ message: 'Товар удален'})
+        } else {
+          return res.status(500).json({ message: 'Не удалось удалить товар'})
+        }
       } else {
-        return res.status(500).json({ message: 'Не удалось удалить товар'})
+        return res.status(500).json({ message: 'Товар не найден'})
       }
     } catch (e) {
       console.log(e);
